+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file
 - Apply theme background colors to mobile & tablet unsupported screen
 - Added `pnpm count` command for code statistics
 - Added `Flashmark` to projects list
+- Added multi-icon support for skill items
 
 ## 0.0.24 (2026-02-05)